Uploading progress on the all in one search filter/ results table
Available at /results_page, or on our instance, the home page. Also tweaked the searchByISSN to allow portions of an ISSN to pick up the full ISSN, rather than requiring the entire thing. May or may not be desired function. Currently the file upload doesn't do anything, and you can only search ONE ISSN at a time, so it's a work in progress.